+++ lib/l10n/app_localizations_es.dart
@@ -9750,12 +9750,6 @@ class AppLocalizationsEs extends AppLocalizations {
   String get settingsWhatsNewTitle => 'Novedades';
 
   @override
-  String get settingThemingDark => 'Tema oscuro';
-
-  @override
-  String get settingThemingLight => 'Tema claro';
-
-  @override
   String get sidebarActiveSectionTitle => 'Actividad';
 
   @override