+++ lib/l10n/app_localizations_fr.dart
@@ -9795,12 +9795,6 @@ class AppLocalizationsFr extends AppLocalizations {
   String get settingsWhatsNewTitle => 'Quoi de neuf';
 
   @override
-  String get settingThemingDark => 'Thème sombre';
-
-  @override
-  String get settingThemingLight => 'Thème clair';
-
-  @override
   String get sidebarActiveSectionTitle => 'Activité';
 
   @override