+++ lib/l10n/app_localizations_pt.dart
@@ -9698,12 +9698,6 @@ class AppLocalizationsPt extends AppLocalizations {
   String get settingsWhatsNewTitle => 'Novidades';
 
   @override
-  String get settingThemingDark => 'Tema escuro';
-
-  @override
-  String get settingThemingLight => 'Tema claro';
-
-  @override
   String get sidebarActiveSectionTitle => 'Atividade';
 
   @override