+++ lib/l10n/app_localizations_ro.dart
@@ -9781,12 +9781,6 @@ class AppLocalizationsRo extends AppLocalizations {
   String get settingsWhatsNewTitle => 'Ce este nou';
 
   @override
-  String get settingThemingDark => 'Temă întunecată';
-
-  @override
-  String get settingThemingLight => 'Temă luminoasă';
-
-  @override
   String get sidebarActiveSectionTitle => 'Activitate';
 
   @override