+++ lib/l10n/app_localizations_sv.dart
@@ -9605,12 +9605,6 @@ class AppLocalizationsSv extends AppLocalizations {
   String get settingsWhatsNewTitle => 'Vad är nytt';
 
   @override
-  String get settingThemingDark => 'Mörkt tema';
-
-  @override
-  String get settingThemingLight => 'Ljustema';
-
-  @override
   String get sidebarActiveSectionTitle => 'Verksamhet';
 
   @override